fix: cursor의 자료형을 int -> Long 타입으로 변경

+++ src/main/java/com/ssafy/ssafsound/domain/post/repository/HotPostCustomRepository.java
@@ -5,5 +5,5 @@ import com.ssafy.ssafsound.domain.post.domain.HotPost;
 import java.util.List;
 
 public interface HotPostCustomRepository {
-    List<HotPost> findWithDetailsFetch(int cursor, int size);
+    List<HotPost> findWithDetailsFetch(Long cursor, int size);
 }
